# perf(autoloader): Drop legacy class autoloader