@ChristophWurst 2023-01-12
> a question also is how difficult it is, whether it's just an hour of work or requires rewriting of things

It's about renaming php class files. The effort depends on the number of files that need a rename.