@nickvergessen 2023-01-12
> How certain are we this is not used in core?

I guess there is only one way to find out: 🍿