@AndyScherzinger 2023-01-13
> This is a hard breaking change for the apps that have not yet been upgraded to PSR-4. Therefore I would suggest to warn about this in the 26 upgrade docs and drop the loader only with Nextcloud 27 or later.

As suggested by Christoph, let's warn with 26 and drop on master after release of v26 that should give everyone enough time to upgrade or fix respectively, including us on core/server if needed.