@ChristophWurst 2023-01-13
> let's warn with 26

PR at https://github.com/nextcloud/documentation/pull/9553