@nickvergessen 2023-05-30
Documentation is in https://github.com/nextcloud/documentation/pull/10231